♻️Refactor : pricing 패키지에 provider 분리, boardId Long 타입으로 수정, board 패키지에 entity, repository 추가

+++ src/main/java/oridungjeol/duckhang/board/infrastructure/entity/SellEntity.java
@@ -11,10 +11,10 @@ import lombok.Setter;
 public class SellEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column(name = "board_id")
-    private int boardId;
+    private Long boardId;
 
     private int price;
